PIXI + Phaser Definitions
Thanks to @natelong for Maintaining the original PIXI definitions.

Updated to 1.5.2 PIXI. 

Needs testing. May be missing updates from the dev branch of Phaser 1.2 that occurred in the last few days.  

Warning! I made Phaser.point extend PIXI.Point. It was the only way I could get this to work. 

If class property position is PIXI.Point
And you try to overlay position with Phaser.Point

I get "X cannot extend Y type mismatch" for a whole bunch of classes. Only when I extend PIXI.Point does the errors dissapear. 

Also was not sure to include PIXI in this file or a separate one.